DFR-2944 Concurrency issue in FinremGeneralOrderRaisedCorresponder (#1646)
* DFR-2945: Refactor GO corresponder

* test fix

* checkstyle fix

* Added tests

* Clean up code

* DFR-2945-update intervener unit test

* Add intervener test to FinremGeneralOrderRaisedContestedCorresponderTest

* Remove redundant classes

---------
